tests/requirements.txt
google-cloud-datastore==1.11.0
requests==2.23.0